+++ next.config.ts
@@ -1,8 +1,9 @@
 import type { NextConfig } from "next";
-
+const basePath = process.env.NEXT_PUBLIC_BASE_PATH;
 const nextConfig: NextConfig = {
   /* config options here */
   reactStrictMode: false,
+  basePath: basePath
 };
 
 export default nextConfig;
