+++ src/app/(secure)/report-builder/util/ReportBuilderUtil.ts
@@ -82,3 +82,5 @@ functions: [
     }
 ];
 
+export const basePath = process.env.NEXT_PUBLIC_BASE_PATH;
+
