+++ src/app/(secure)/report-types/define-relationships/page.tsx
@@ -1,6 +1,12 @@
-"use client";
+'use client';
 import DefineRelationshipsForm from "@/components/report-type/define-relationship-form";
+import { Suspense } from "react";
 
 export default function SelectRelationshipPage() {
-  return <DefineRelationshipsForm reportTypeId={""} ></DefineRelationshipsForm>
+
+  return (
+      <Suspense fallback={<div>Loading...</div>}>
+        <DefineRelationshipsForm reportTypeId={""} ></DefineRelationshipsForm>
+      </Suspense>
+    );
 }
