+++ src/app/(secure)/report-types/select-object/[reportTypeId]/page.tsx
@@ -1,6 +1,12 @@
+'use client';
 import SelectObjectForm from "@/components/report-type/select-object-form";
+import { Suspense } from "react";
 
 export default function ReportTypeSelectionPage({ params }: { params: any }) {
-    console.log("Edit report type page ------------------------");
-    return <SelectObjectForm reportTypeId={params?.reportTypeId} ></SelectObjectForm>;
+
+    return (
+        <Suspense fallback={<div>Loading...</div>}>
+            <SelectObjectForm reportTypeId={params?.reportTypeId} ></SelectObjectForm>;
+        </Suspense>
+    );
 }
