+++ src/app/(secure)/report-types/select-object/page.tsx
@@ -1,6 +1,9 @@
+'use client';
 import SelectObjectForm from "@/components/report-type/select-object-form";
+import { Suspense } from "react";
 
 export default function ReportTypeSelectionPage() {
-  console.log("Create report type page ------------------------");
-  return <SelectObjectForm reportTypeId="" ></SelectObjectForm>
+  return <Suspense fallback={<div>Loading...</div>}>
+    <SelectObjectForm reportTypeId="" ></SelectObjectForm>
+  </Suspense>
 }
