+++ src/app/(secure)/report-types/select-relationships/[reportTypeId]/page.tsx
@@ -1,6 +1,11 @@
 "use client";
 import SelectRelationshipsForm from "@/components/report-type/select-relationship-form";
+import { Suspense } from "react";
 
-export default function SelectRelationshipPage({params}: {params: any}) {
-  return <SelectRelationshipsForm reportTypeId={params.reportTypeId} ></SelectRelationshipsForm>
+export default function SelectRelationshipPage({ params }: { params: any }) {
+  return (
+    <Suspense fallback={<div>Loading...</div>}>
+      <SelectRelationshipsForm reportTypeId={params.reportTypeId} />
+    </Suspense>
+  );
 }
