+++ src/app/(secure)/report-types/select-relationships/page.tsx
@@ -2,6 +2,5 @@
 import SelectRelationshipsForm from "@/components/report-type/select-relationship-form";
 
 export default function SelectRelationshipPage() {
-  console.log("Create SelectRelationshipsForm page ------------------------");
   return <SelectRelationshipsForm reportTypeId="" ></SelectRelationshipsForm>
 }
