+++ src/components/charts/GroupedBarChart.tsx
@@ -20,7 +20,7 @@ const data = [
   { name: 'Jun', group1: 700, group2: 800 },
 ];
 
-export function GroupedBarChart() {
+export function GroupedBarChart({ data, config }) {
   return (
     <ResponsiveContainer width="100%" height={300}>
       <BarChart data={data}>
